+++ inc/simterms.php
@@ -68,6 +68,10 @@ function getSimilarityRanking($str1, $str2)
 /**
  * For a language $lang_id and a term $compared_term (UTF-8).
  * If string is already in database, it will be excluded in results.
+ * 
+ * @param int $lang_id Language ID
+ * @param int max_count Maximum number of terms to display
+ * 
  *
  * @return int[] All $max_count wordids with a similarity ranking > $min_ranking, 
  *               sorted decending
@@ -157,6 +161,10 @@ function format_term($termid, $compare)
 /**
  * Get Term and translation of terms in termid array (calculated 
  * in function get_similar_terms(...)) as string for echo
+ * 
+ * @param int    $lang_id       Language ID
+ * @param string $compared_term Similar term we compare to
+ * 
  */
 function print_similar_terms($lang_id, $compared_term): string 
 {